+++ packages/store/src/internal/dispatcher.ts
@@ -64,12 +64,14 @@ export class InternalDispatcher {
   }
 
   private dispatchSingle(action: any): Observable<any> {
-    const type: string | undefined = getActionTypeFromInstance(action);
-    if (!type) {
-      const error = new Error(
-        `This action doesn't have a type property: ${action.constructor.name}`
-      );
-      return throwError(error);
+    if (typeof ngDevMode === 'undefined' || ngDevMode) {
+      const type: string | undefined = getActionTypeFromInstance(action);
+      if (!type) {
+        const error = new Error(
+          `This action doesn't have a type property: ${action.constructor.name}`
+        );
+        return throwError(error);
+      }
     }
 
     const prevState = this._stateStream.getValue();