+++ packages/store/src/module.ts
@@ -20,7 +20,7 @@ import {
   NgxsModuleOptions,
   ROOT_STATE_TOKEN
 } from './symbols';
-import { NGXS_EXECUTION_STRATEGY } from './execution/symbols';
+import { USER_PROVIDED_NGXS_EXECUTION_STRATEGY } from './execution/symbols';
 import { StateFactory } from './internal/state-factory';
 import { StateContextFactory } from './internal/state-context-factory';
 import { Actions, InternalActions } from './actions-stream';
@@ -28,11 +28,11 @@ import { LifecycleStateManager } from './internal/lifecycle-state-manager';
 import { InternalDispatchedActionResults, InternalDispatcher } from './internal/dispatcher';
 import { InternalStateOperations } from './internal/state-operations';
 import { Store } from './store';
+import { SelectFactory } from './decorators/select/select-factory';
 import { StateStream } from './internal/state-stream';
 import { PluginManager } from './plugin-manager';
 import { NgxsRootModule } from './modules/ngxs-root.module';
 import { NgxsFeatureModule } from './modules/ngxs-feature.module';
-import { DispatchOutsideZoneNgxsExecutionStrategy } from './execution/dispatch-outside-zone-ngxs-execution-strategy';
 import { InternalNgxsExecutionStrategy } from './execution/internal-ngxs-execution-strategy';
 import { mergeDeep } from './utils/utils';
 
@@ -65,6 +65,7 @@ export class NgxsModule {
         InternalNgxsExecutionStrategy,
         Store,
         StateStream,
+        SelectFactory,
         PluginManager,
         ...states,
         ...NgxsModule.ngxsTokenProviders(states, options)
@@ -97,8 +98,8 @@ export class NgxsModule {
   ): Provider[] {
     return [
       {
-        provide: NGXS_EXECUTION_STRATEGY,
-        useClass: options.executionStrategy || DispatchOutsideZoneNgxsExecutionStrategy
+        provide: USER_PROVIDED_NGXS_EXECUTION_STRATEGY,
+        useValue: options.executionStrategy
       },
       {
         provide: ROOT_STATE_TOKEN,